Books can now be added to accounts
