+++ app/controllers/accounts/base_controller.rb
@@ -1,0 +1,15 @@
+module Accounts
+  class BaseController < ApplicationController
+    before_action :authenticate_user!
+    
+    def current_account
+      @current_account ||= Account.find_by!(subdomain: request.subdomain)
+    end
+    helper_method :current_account
+    
+    def owner?
+      current_account.owner == current_user
+    end
+    helper_method :owner?
+  end
+end
