+++ app/views/accounts/chapters/_ipad.html.erb
