+++ app/views/accounts/elements/_footnote.html.erb
