+++ app/views/accounts/elements/_img.html.erb
@@ -2,7 +2,7 @@
 <% image = @chapter.images.find_by(filename: image_html["src"]) %>
 
 <% if show_notes %>
-  <%= render "notes/button", element: element %>
+  <%= render "accounts/notes/button", element: element %>
 <% end %>
 
 <div class='image'>
