+++ app/views/accounts/notes/_note.html.erb
