+++ app/views/accounts/notes/show.html.erb
@@ -14,6 +14,6 @@
       <%= render @comments.select(&:persisted?) %>
     </div>
 
-    <%= render "comments/form" %>
+    <%= render "accounts/comments/form" %>
   </div>
 </div>
