+++ lib/constraints/subdomain_required.rb
@@ -1,0 +1,5 @@
+class SubdomainRequired
+  def self.matches?(request)
+    request.subdomain.present? && request.subdomain != "www"
+  end
+end
