+++ spec/features/notes_spec.rb
@@ -1,10 +1,11 @@
 require 'rails_helper'
 
 describe "notes" do
-  let(:author) { create_author! }
+  let(:account) { FactoryBot.create(:account) }
   let(:book) { create_book! }
   before do
-    login_as(author)
+    login_as(account.owner)
+    set_subdomain(account.subdomain)
   end
     
   it "can add a new note to a paragraph", :js => true do
@@ -21,7 +22,7 @@ describe "notes" do
     expect(page).to have_content("1 note +")
     click_link "All notes for this chapter"
     click_link "This is a test note!"
-    expect(page).to have_content("#{author.email} commented less than a minute ago")
+    expect(page).to have_content("#{account.owner.email} commented less than a minute ago")
     # Ensure note text shows up correctly in processed markdown.
 
     within ".comment" do
@@ -39,18 +40,18 @@ describe "notes" do
     chapter = book.chapters.first
     element = chapter.elements.first
     note = element.notes.create!(
-      user: author, 
+      user: account.owner, 
       number: 1
     )
     note.comments.create!(
-      user: author,
+      user: account.owner,
       text: "This is a test note!"
     )
     
     visit book_path(book)
     click_link "All notes for this book"
     click_link "This is a test note!"
-    expect(page).to have_content("#{author.email} commented less than a minute ago")
+    expect(page).to have_content("#{account.owner.email} commented less than a minute ago")
     expect(page).to have_content("This is a test note!")
   end
 
@@ -59,12 +60,12 @@ describe "notes" do
       chapter = book.chapters.first
       element = chapter.elements.first
       @note = element.notes.create!(
-        user: author, 
+        user: account.owner, 
         number: 1
       )
       @note.comments.create!(
         text: "This is a test note!",
-        user: author
+        user: account.owner
       )
       
       visit book_path(book)
@@ -90,12 +91,12 @@ describe "notes" do
     chapter = book.chapters.first
     element = chapter.elements.first
     note = element.notes.create!(
-      user: author, 
+      user: account.owner, 
       number: 1,
       state: "rejected"
     )
     note.comments.create!(
-      user: author,
+      user: account.owner,
       text: "This is a test note!"
     )
