+++ spec/support/factories/account_factory.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :account do
+    sequence(:name) { |n| "Test Account ##{n}" }
+    sequence(:subdomain) { |n| "test#{n}" }
+    association :owner, :factory => :user
+  end
+end
